@@ -45,7 +45,6 @@
     "postcss-loader": "^8.1.1",
     "prop-types": "^15.8.1",
     "raw-loader": "^4.0.2",
-    "redux-devtools": "^3.5.0",
     "sass": "^1.89.0",
     "sass-loader": "^16.0.5",
     "style-loader": "^4.0.0",
@@ -72,10 +71,10 @@
     "formik": "^2.4.6",
     "initials": "^3.0.1",
     "prettier": "^3.5.3",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^7.6.1",
+    "react-router-dom": "^7.6.1",
     "webpack-node-externals": "^1.7.2",
     "yup": "^1.6.1"
   }
